SISRP-38270 - Moves updateTaskLists method and all dependent methods to tasksService
